+++ ouroboros-network-framework/src/Ouroboros/Network/Socket.hs
@@ -377,7 +377,6 @@ connectToNodeWithMux'
     connectionId <- (\localAddress remoteAddress -> ConnectionId { localAddress, remoteAddress })
                 <$> Snocket.getLocalAddr sn sd <*> Snocket.getRemoteAddr sn sd
     muxTracer <- initDeltaQTracer' $ Mx.WithBearer connectionId `contramap` nctMuxTracer
-    ts_start <- getMonotonicTime
 
     handshakeBearer <- Mx.getBearer makeBearer sduHandshakeTimeout muxTracer sd Nothing
     app_e <-
@@ -394,18 +393,14 @@ connectToNodeWithMux'
           haTimeLimits       = handshakeTimeLimits
         }
         versions
-    ts_end <- getMonotonicTime
     case app_e of
        Left (HandshakeProtocolLimit err) -> do
-         traceWith muxTracer $ Mx.TraceHandshakeClientError err (diffTime ts_end ts_start)
          throwIO err
 
        Left (HandshakeProtocolError err) -> do
-         traceWith muxTracer $ Mx.TraceHandshakeClientError err (diffTime ts_end ts_start)
          throwIO err
 
        Right (HandshakeNegotiationResult app versionNumber agreedOptions) -> do
-         traceWith muxTracer $ Mx.TraceHandshakeClientEnd (diffTime ts_end ts_start)
          Mx.withReadBufferIO (\buffer -> do
              bearer <- Mx.getBearer makeBearer sduTimeout muxTracer sd buffer
              mux <- Mx.new (toMiniProtocolInfos (runForkPolicy noBindForkPolicy remoteAddress) app)
@@ -414,7 +409,6 @@ connectToNodeWithMux'
            )
 
        Right (HandshakeQueryResult _vMap) -> do
-         traceWith muxTracer $ Mx.TraceHandshakeClientEnd (diffTime ts_end ts_start)
          throwIO (QueryNotSupported @vNumber)